// LiveTalk-main/types.ts
export enum UserLevel {
  NEW = 'جديد',
  BRONZE = 'برونزي',
  SILVER = 'فضي',
  GOLD = 'ذهبي',
  DIAMOND = 'ماسي',
  VIP = 'VIP'
}

export type ItemType = 'frame' | 'bubble' | 'entry' | 'badge' | 'extra';

export interface StoreItem {
  id: string;
  name: string;
  type: ItemType;
  price: number;
  url: string;
}

export interface CPPartner {
  id: string;
  name: string;
  avatar: string;
  type: 'cp' | 'friend';
  level?: number;
}

export interface HostAgency {
  id: string;
  name: string;
  agentId: string;
  agentName: string;
  createdAt: any;
  totalProduction: number;
}

export interface User {
  id: string;
  customId: any; 
  name: string;
  avatar: string;
  level: UserLevel;
  wealthLevel?: number;
  rechargeLevel?: number;
  rechargePoints?: number;
  frame?: string;
  activeBubble?: string;
  badge?: string;
  achievements?: string[];
  cover?: string;
  coins: any;
  diamonds: any; 
  wealth: any;
  charm: any;
  isVip: boolean;
  vipLevel?: number;
  nameStyle?: string;
  idColor?: string; 
  bio?: string;
  location?: string;
  gender?: 'male' | 'female';
  stats?: {
    likes: number;
    visitors: number;
    following: number;
    followers: number;
  };
  ownedItems?: string[];
  isFollowing?: boolean;
  isMuted?: boolean;
  isSpecialId?: boolean;
  isAdmin?: boolean;
  isAgency?: boolean; // وكيل شحن
  agencyBalance?: number;
  isHostAgent?: boolean; // وكيل مضيفين
  hostAgencyId?: string; // لربط المضيف بوكالة
  isHost?: boolean; // هل هو مضيف مسجل
  hostProduction?: number; // تارجت المضيف
  isBanned?: boolean;
  banUntil?: string;
  seatIndex?: number;
  status?: string;
  activeEmoji?: string; 
  cpPartner?: CPPartner | null;
  friendPartner?: CPPartner | null; // حقل جديد للصداقة المقربة
  roomTemplate?: {
    title: string;
    category: string;
    thumbnail: string;
    background: string;
    isLocked: boolean;
    password?: string;
  };
}